requirements.txt
google-genai
pydantic

httpx
requests
beautifulsoup4
